rename React component files
