+++ client/components/index.js
@@ -3,10 +3,10 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Main} from './main';
-export {default as UserHome} from './user-home';
-export {default as UploadActivity} from './upload-activity';
+export {default as Main} from './Main';
+export {default as UserHome} from './UserHome';
+export {default as UploadActivity} from './UploadActivity';
 export {default as SingleActivity} from './SingleActivity';
 export {default as AllActivities} from './AllActivities';
-export {Login, Signup} from './auth-form';
+export {Login, Signup} from './AuthForm';
